src/generator/question_generator.py
from langchain_core.output_parsers import PydanticOutputParser
from src.models.question_schemas import MCQQuestion, FillBlankQuestion
from src.prompts.templates import mcq_prompt_template, fill_blank_prompt_template
from src.llm.gorq_client import get_groq_llm
from src.config.settings import Settings
from src.utils.logger import get_logger
from src.utils.custom_exception import CustomException

class QuestionGenerator:
  def __init__(self):
    self.llm = get_groq_llm()
    self.logger = get_logger(self.__class__.__name__)
  
  def _retry_and_parse(self, prompt, parser, topic, difficulty):
    for attempt in range(Settings.MAX_RETRIES):
      try:
        self.logger.info(f"Generate question with {topic} with difficulty '{difficulty}' ")
        response = self.llm.invoke(prompt.format(topic=topic, difficulty=difficulty))
        parsed = parser.parse(response.content)
        self.logger.info("SUCCESS parsed the question")
        return parsed
      except Exception as e:
        self.logger.error(f"{__name__}, {str(e)}")
        if attempt==Settings.MAX_RETRIES-1:
          raise CustomException(f"Generation failed afer {Settings.MAX_RETRIES} attempts")
  
  def generate_mcq(self, topic: str, difficulty: str ='medium') -> MCQQuestion:
    try:
      parser = PydanticOutputParser(pydantic_object=MCQQuestion)
      question = self._retry_and_parse(mcq_prompt_template, parser, topic, difficulty)
      if len(question.options) != 4 or question.correct_answer not in question.options:
        raise ValueError("Invalid MCQ structure")
      self.logger.info("Generated a valid MCQ")
      return question
    except Exception as e:
      self.logger.error(f"Failed to generate MCQ, {str(e)}")
      raise CustomException(f"MCQ question failed {e}")

  def generate_fill_blank(self, topic: str, difficulty: str ='medium') -> FillBlankQuestion:
    try:
      parser = PydanticOutputParser(pydantic_object=FillBlankQuestion)
      question = self._retry_and_parse(fill_blank_prompt_template, parser, topic, difficulty)
      if "___" not in question.question:
        raise ValueError("Question should have '___'")
      self.logger.info("Generated a valid fill in blanks")
      return question
    except Exception as e:
      self.logger.error(f"Failed to generate fill in blanks, {str(e)}")
      raise CustomException(f"fill in blanks question failed {e}")
