requirements.txt
langchain
langchain-huggingface
langchain-groq
langchain-community
python-dotenv
black
pandas
flask
setuptools
streamlit